+++ azure-cli-diff-tool/azureCliDiffTool/meta_change.py
@@ -4,11 +4,9 @@
 # license information.
 # -----------------------------------------------------------------------------
 
-# pylint: disable=too-many-instance-attributes
-
-from azdev.utilities import get_change_rule_template, get_change_suggest_template
-from azdev.utilities import CMD_PROPERTY_IGNORED_LIST, PARA_PROPERTY_IGNORED_LIST, PARA_NAME_IGNORED_LIST, \
-    PARA_VALUE_IGNORED_LIST, BREAKING_CHANE_RULE_LINK_URL_PREFIX, BREAKING_CHANE_RULE_LINK_URL_SUFFIX
+from .utils import get_change_rule_template, get_change_suggest_template
+from ._const import BREAKING_CHANE_RULE_LINK_URL_PREFIX, BREAKING_CHANE_RULE_LINK_URL_SUFFIX, \
+    CMD_PROPERTY_IGNORED_LIST, PARA_NAME_IGNORED_LIST, PARA_PROPERTY_IGNORED_LIST, PARA_VALUE_IGNORED_LIST
 
 
 class MetaChange:
@@ -16,10 +14,10 @@ class MetaChange:
     def __init__(self, rule_id="1000", is_break=False, rule_message="", suggest_message="",
                  is_ignore=False, filter_key=None):
         self.rule_id = rule_id
-        self.rule_link_url = BREAKING_CHANE_RULE_LINK_URL_PREFIX + self.rule_id + BREAKING_CHANE_RULE_LINK_URL_SUFFIX
         self.is_break = is_break
         self.rule_message = rule_message
         self.suggest_message = suggest_message
+        self.rule_link_url = BREAKING_CHANE_RULE_LINK_URL_PREFIX + self.rule_id + BREAKING_CHANE_RULE_LINK_URL_SUFFIX
         self.is_ignore = is_ignore
         self.filter_key = filter_key