+++ HISTORY.rst
@@ -2,6 +2,18 @@
 
 Release History
 ===============
+0.1.55
+++++++
+* `azdev command-change meta-diff`: Fix data/meta_change_whitelist.txt pkg building
+
+0.1.54
+++++++
+* `azdev command-change meta-export/meta-diff`: Add cmd `type` extraction, fix false positive in cmdPropUpdate and add whitelist in certain meta-changes (#401)
+
+0.1.53
+++++++
+* `azdev command-change meta-export`: Add rule link (#402)
+
 0.1.52
 ++++++
 * Fix cmdcov issue(#396): Fix the regex which to get the command group